+++ src/main/java/org/example/pdf_creator/content/PdfCreationConfiguration.java
@@ -2,27 +2,19 @@ package org.example.pdf_creator.content;
 
 import com.itextpdf.layout.element.BlockElement;
 import lombok.Data;
-import lombok.NoArgsConstructor;
 import org.example.pdf_creator.content.abstractsclasses.TextSection;
 import org.example.pdf_creator.content.abstractsclasses.TextSectionList;
 import org.example.pdf_creator.factories.TextSectionListType;
 
-import java.util.ArrayList;
-import java.util.Collections;
 import java.util.List;
 
 @Data
 public class PdfCreationConfiguration extends TextSectionList {
 
-    public PdfCreationConfiguration(List<TextSection> sections) {
-        super(sections);
-        if(this.textSectionList.size() > 1) {
-            Collections.sort(this.textSectionList);
-        }
+    public PdfCreationConfiguration(List<TextSection> textSectionList, List<FontPreset> fontPresets) {
+        super(textSectionList, fontPresets);
     }
-    public PdfCreationConfiguration() {
-        this(new ArrayList<>());
-    }
+
     @Override
     public BlockElement createContent() {
         return null;